+++ tests/unit/test_main.py
@@ -1,7 +1,5 @@
 import logging
-import os
 
-import pytest
 import yaml
 from click.testing import CliRunner