+++ tests/unit/test_scan.py
@@ -40,8 +40,7 @@ class TestScan:
     class TestWhenCouldNotFindAPISpecFile:
         def test_should_log_error(self, mocker, caplog):
             mocker.patch(
-                "scanapi.scan.settings",
-                {"spec_path": "invalid_path/scanapi.yaml", "no_report": False},
+                "scanapi.scan.settings", {"spec_path": "invalid_path/scanapi.yaml"}
             )
             mocker.patch("scanapi.scan.load_config_file", side_effect=file_not_found)
             with caplog.at_level(logging.ERROR):
@@ -143,7 +142,7 @@ class TestWriteReporter:
         mock_write = mocker.patch("scanapi.scan.Reporter.write")
         mock_reporter_init = mocker.patch("scanapi.scan.Reporter.__init__")
         mock_reporter_init.return_value = None
-        mock_settings = mocker.patch(
+        mocker.patch(
             "scanapi.scan.settings",
             {
                 "output_path": "out/my-report.md",