+++ tests/unit/tree/test_request_node.py
@@ -22,7 +22,7 @@ class TestRequestNode:
 
         def test_missing_required_keys(self):
             with pytest.raises(MissingMandatoryKeyError) as excinfo:
-                request = RequestNode(
+                RequestNode(
                     spec={}, endpoint=EndpointNode({"name": "foo", "requests": [{}]}),
                 )
 
@@ -67,9 +67,7 @@ class TestRequestNode:
         @pytest.mark.skip("it should validate mandatory `name` key before")
         def test_when_request_has_no_name(self):
             with pytest.raises(MissingMandatoryKeyError) as excinfo:
-                request = RequestNode(
-                    {}, endpoint=EndpointNode({"name": "foo", "requests": []})
-                )
+                RequestNode({}, endpoint=EndpointNode({"name": "foo", "requests": []}))
 
             assert str(excinfo.value) == "Missing name, path at 'request'"
 
@@ -88,7 +86,7 @@ class TestRequestNode:
                 {"name": "foo", "requests": [{}], "path": endpoint_path}
             )
             request = RequestNode({"path": "/foo", "name": "foo"}, endpoint=endpoint)
-            assert request.full_url_path == f"http://foo.com/api/foo"
+            assert request.full_url_path == "http://foo.com/api/foo"
 
         def test_with_trailing_slashes(self):
             endpoint = EndpointNode(
@@ -129,7 +127,7 @@ class TestRequestNode:
             request = RequestNode(
                 {"headers": headers, "path": "http://foo.com", "name": "foo"},
                 endpoint=EndpointNode(
-                    {"headers": endpoint_headers, "name": "foo", "requests": [{}],}
+                    {"headers": endpoint_headers, "name": "foo", "requests": [{}]}
                 ),
             )
             assert request.headers == {"abc": "def", "xxx": "www"}
@@ -140,7 +138,7 @@ class TestRequestNode:
             request = RequestNode(
                 {"headers": headers, "path": "http://foo.com", "name": "foo"},
                 endpoint=EndpointNode(
-                    {"headers": endpoint_headers, "name": "foo", "requests": [{}],}
+                    {"headers": endpoint_headers, "name": "foo", "requests": [{}]}
                 ),
             )
             assert request.headers == {"abc": "def", "xxx": "www"}
@@ -151,7 +149,7 @@ class TestRequestNode:
             )
 
             request = RequestNode(
-                {"headers": {"ghi": "jkl"}, "path": "http://foo.com", "name": "foo",},
+                {"headers": {"ghi": "jkl"}, "path": "http://foo.com", "name": "foo"},
                 endpoint=endpoint,
             )
             request.headers
@@ -196,7 +194,7 @@ class TestRequestNode:
             )
 
             request = RequestNode(
-                {"params": {"ghi": "jkl"}, "path": "http://foo.com", "name": "foo",},
+                {"params": {"ghi": "jkl"}, "path": "http://foo.com", "name": "foo"},
                 endpoint=endpoint,
             )
             request.params
@@ -240,14 +238,14 @@ class TestRequestNode:
 
         def test_when_request_has_body(self):
             request = RequestNode(
-                {"body": {"abc": "def"}, "path": "http://foo.com", "name": "foo",},
+                {"body": {"abc": "def"}, "path": "http://foo.com", "name": "foo"},
                 endpoint=EndpointNode({"name": "foo", "requests": [{}]}),
             )
             assert request.body == {"abc": "def"}
 
         def test_calls_evaluate(self, mocker, mock_evaluate):
             request = RequestNode(
-                {"body": {"ghi": "jkl"}, "path": "http://foo.com", "name": "foo",},
+                {"body": {"ghi": "jkl"}, "path": "http://foo.com", "name": "foo"},
                 endpoint=EndpointNode({"name": "foo", "requests": [{}]}),
             )
             request.body