+++ CHANGELOG.md
@@ -6,11 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- Add `--no-report` flag. [#299](https://github.com/scanapi/scanapi/pull/299)
-
 ### Changed
 - Updated poetry-publish version to v1.3 [#311](https://github.com/scanapi/scanapi/pull/311)
+
+### Fixed
+- Removed all flake8 errors [#309](https://github.com/scanapi/scanapi/pull/309)
 
 ## [2.1.0] - 2020-10-06
 ### Added